Log: controller: add FSA action flag functions
... instead of set_bit()/clear_bit(), which improves trace messages like:

    Bit 0x00000200 set by some_function:100

to:

    FSA action flags 0x00000200 (A_DC_TIMER_STOP) for controller set by some_function:100

The new flag functions also replace the register_fsa_action() wrapper. To keep
similar behavior, they log at debug level, rather than the usual trace.

+++ daemons/controld/controld_matrix.h
@@ -1,19 +1,10 @@
 /*
- * Copyright (C) 2004 Andrew Beekhof <andrew@beekhof.net>
+ * Copyright 2004-2020 the Pacemaker project contributors
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2 of the License, or (at your option) any later version.
+ * The version control history for this file may have further details.
  *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ * This source code is licensed under the GNU Lesser General Public License
+ * version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
  */
 
 #ifndef XML_FSA_MATRIX__H
@@ -627,7 +618,7 @@ const enum crmd_fsa_state crmd_fsa_state[MAXINPUT][MAXSTATE] = {
 
 /* NOTE: In the fsa, the actions are extracted then state is updated. */
 
-const long long crmd_fsa_actions[MAXINPUT][MAXSTATE] = {
+const uint64_t crmd_fsa_actions[MAXINPUT][MAXSTATE] = {
 
 /* Got an I_NULL */
     {
